+++ CHANGELOG.md
@@ -14,4 +14,5 @@ All notable changes to this project will be documented in this file. See [standa
 * [#19](https://github.com/upfrontsystems/izinto-angular/issues/19) add ordering to chart
 * [#20](https://github.com/upfrontsystems/izinto-angular/issues/19) add data source page
 * [#29](https://github.com/upfrontsystems/izinto-angular/issues/29) fix chart reordering and UI
+* [#30](https://github.com/upfrontsystems/izinto-angular/issues/30) handle start and end range query